[autolinking] Restructure code
