+++ packages/expo-modules-autolinking/src/autolinking/index.ts
@@ -1,0 +1,5 @@
+export { findModulesAsync } from './findModules';
+export { generatePackageListAsync } from './generatePackageList';
+export { mergeLinkingOptionsAsync } from './mergeLinkingOptions';
+export { resolveModulesAsync } from './resolveModules';
+export { verifySearchResults } from './verifySearchResults';
